Should have a new section here
Based on how the rest of this section is written, the sentence starting with "In some cases..." should have its own section, something like "Operational Guidance".

This is a suggestion for how to do it. Without this though the previous section on Data Authentication goes right into this without a break and it makes the references confusion (since you are still looking at Data Authentication but you are now onto a different topic without a note to the change).